+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/stop/BikeRentalStationType.java
@@ -6,7 +6,7 @@ import graphql.schema.GraphQLInterfaceType;
 import graphql.schema.GraphQLList;
 import graphql.schema.GraphQLNonNull;
 import graphql.schema.GraphQLObjectType;
-import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+import org.opentripplanner.routing.vehicle_rental.BikeRentalStation;
 
 import java.util.ArrayList;
