+++ src/main/java/org/opentripplanner/api/mapping/VehicleRentalStationMapper.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.api.mapping;
 
 import org.opentripplanner.api.model.ApiVehicleRentalStation;
-import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+import org.opentripplanner.routing.vehicle_rental.BikeRentalStation;
 
 import java.util.ArrayList;
 import java.util.Locale;
