+++ src/main/java/org/opentripplanner/routing/vehicle_rental/BikeRentalStationService.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.bike_rental;
+package org.opentripplanner.routing.vehicle_rental;
 
 import java.io.Serializable;
 import java.util.Collection;
