+++ src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/GenericJsonVehicleRentalDataSource.java
@@ -3,7 +3,7 @@ package org.opentripplanner.updater.vehicle_rental.datasources;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+import org.opentripplanner.routing.vehicle_rental.BikeRentalStation;
 import org.opentripplanner.updater.vehicle_rental.VehicleRentalDataSource;
 import org.opentripplanner.updater.vehicle_rental.datasources.params.VehicleRentalDataSourceParameters;
 import org.opentripplanner.util.HttpUtils;
